requirements.txt
selenium==4.13.0
pytest==7.4.2
pytest-html==4.0.2
pytest-rerunfailures==12.0
pytest-sugar==0.9.7
pytest-xdist==3.3.1
requests==2.31.0
boto3==1.28.57
furl~=2.1.0
behave==1.2.6
colorama==0.4.6
pyyaml==6.0.1
webdriver_manager==4.0.1
python-dateutil~=2.8.1
DateTime~=5.2
pandas~=2.1.1
openpyxl~=3.1.2
behave2cucumber~=1.0.3
axe-selenium-python~=2.1.6
behave-html-formatter~=0.9.10
